fix: file an issue when the milestone isn't found
If the repo doesn't contain the milestone, we still need to
create the issue. This prevents losing the results of the test
run. It also may handle the case where the repo is forked and the
milestone is missing from the fork.

This means that we make a best-effort attempt to attach a
milestone to the issue. But if we cannot find one, we file the
issue with no milestone linkage.